move mutation-invariance test to test_trace_server_interface_util
test_trace_server_conftest.py is for testing the fixture machinery,
wrong home for an adapter unit test. test_trace_server_interface_util
already holds trace-server-interface-level unit tests.
